+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This is the changelog for [OpenAPI Factory](readme.md).
 
 ## 1.1 ##
 * Convert to using singlu function to Lambda Functions will work with the API without class instantiation.
+* Add VPC Security Group and Subnet options.
 
 ## 1.0 ##
 * Created npm library for building apis.